requirements.txt
# Data Processing and ETL
pandas>=1.5.0
numpy>=1.21.0
snowflake-connector-python>=3.0.0
snowflake-sqlalchemy>=1.4.0
sqlalchemy>=1.4.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.1.0

# AI and NLP
langchain>=0.0.200
openai>=0.27.0
python-dotenv>=0.19.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.65.0
pytest>=7.0.0
jupyter>=1.0.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.13.0

# API and Web
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0 